Moved interfaces to Core
